- Modal extracted to hook
